+++ vox/Source/vox/src/Texture.cpp
@@ -5,7 +5,7 @@
 #include <GLUtils.h>
 #include <SOIL.h>
 
-Texture::Texture(const char* _src, int _width, int _height, bool _storeData, bool _autoRelease) :
+Texture::Texture(const char* _src, unsigned long int _width, unsigned long int _height, bool _storeData, bool _autoRelease) :
 	NodeResource(_autoRelease),
 	src(_src),
 	width(_width),
